+++ CHANGES.md
@@ -31,6 +31,7 @@
   `fullscreen` and `mobile` respectively.
 - Fetch VCard when starting a chat with someone not in the user's roster.
 - Show status messages in an MUC room when a user's role changes.
+- In MUC chat rooms, collapse multiple, consecutive join/leave messages.
 
 ### API changes
 - New API method `_converse.disco.supports` to check whether a certain